chore: Move fonts to public
