+++ components/layouts/Layout.js
@@ -2,7 +2,7 @@ import Head from "next/head";
 import Menu from "@/components/Menu/Menu";
 import Footer from "../Footer";
 
-import { skModernist } from "@/src/fonts/font-provider";
+import { skModernist } from "@/src/public/fonts/font-provider";
 
 const Layout = ({ children, router }) => {
   return (
